requirements.txt
hazm==0.7.0
Python 3.8.13
wordcloud==1.8.2.2
python-bidi==0.4.2
arabic-reshaper==2.1.3